nu_plugin_secret

Module tera_functions

Tera template functions for redaction templates

This module provides custom Tera template functions that can be used in redaction templates for flexible secret formatting.

Available functions:

replicate(s="*", n=5): Returns a string of repeated characters

reverse(s="text"): Returns the input string reversed

take(n=5, s="text"): Returns the first n characters of input string

strlen(s="text"): Returns the length of the input string as a number

mask_partial(s="text", l=2, r=2): Masks the middle portion of a string, keeping specified characters from left and right (l and r default to 0, optional c=“*” for custom masking character)

secret_string(): Returns the actual secret value (WARNING: exposes sensitive data!)

Functions

register_all_standard_functions

Register all standard template functions (excluding secret_string)

register_mask_partial_function

Register the mask_partial function with a Tera instance Masks the middle portion of a string, keeping specified characters from left and right

register_replicate_function

Register the replicate function with a Tera instance Returns a string of the given character repeated length times

register_reverse_function

Register the reverse function with a Tera instance Returns the input string reversed

register_strlen_function

Register the strlen function with a Tera instance Returns the length of the input string as a number

register_take_function

Register the take function with a Tera instance Returns the first n characters of the input string